requirements.txt
Flask==2.1.2
flask-restx==0.5.1
Flask-Migrate==3.1.0
python-dotenv==0.20.0
requests==2.28.1
celery==5.2.7
flask-sqlalchemy==2.5.1
flask-script==2.0.6
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
flask-limiter==1.1.0
gunicorn==20.1.0
Jinja2==3.1.2
markupsafe==2.1.1
itsdangerous==2.1.2
werkzeug==2.1.2
flask-caching==1.11.1
cachelib==0.8.0
sentry-sdk==1.17.0
SQLAlchemy==1.4.37
blinker==1.5
pymysql==1.0.2
gunicorn==20.1.0
redis==4.3.3
flower==1.2.0
flask-cors==3.0.10
flask-security-too==5.1.2
flask-admin==1.6.1